flask/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Cors